Move files to more correct location
