+++ src/app/features/call/SmallWidget.ts
